refactor: move sde_window_scheduler to pipelines/schedulers
